requirements.txt
langchain
openai
streamlit
python-dotenv
requests
